requirements.txt
fastapi
uvicorn
python-dotenv
groq
langchain
langchain-text-splitters
pinecone-client
pypdf
requests